+ collect javaDev4
Java Dev 4 was, probably, one of the toughest assignment Mr Yu gave to
us. There was, honestly, no one else wanted to spend time resarching on
it, except for myself. At the end of the semester when Mr Yu checked
homework, uh.. report to be precise, everyone else copied mine. Was
exciting and proud of it but didn't think that Java is never my cup of
tea.